[Improvement][api] Introduce executor and datasource service interface for clear code
