+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/ExecutorService2Test.java
@@ -16,17 +16,36 @@
  */
 package org.apache.dolphinscheduler.api.service;
 
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
 import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.service.impl.ExecutorServiceImpl;
 import org.apache.dolphinscheduler.common.Constants;
 import org.apache.dolphinscheduler.common.enums.CommandType;
 import org.apache.dolphinscheduler.common.enums.Priority;
 import org.apache.dolphinscheduler.common.enums.ReleaseState;
 import org.apache.dolphinscheduler.common.enums.RunMode;
 import org.apache.dolphinscheduler.common.model.Server;
-import org.apache.dolphinscheduler.dao.entity.*;
+import org.apache.dolphinscheduler.dao.entity.Command;
+import org.apache.dolphinscheduler.dao.entity.ProcessDefinition;
+import org.apache.dolphinscheduler.dao.entity.Project;
+import org.apache.dolphinscheduler.dao.entity.Schedule;
+import org.apache.dolphinscheduler.dao.entity.Tenant;
+import org.apache.dolphinscheduler.dao.entity.User;
 import org.apache.dolphinscheduler.dao.mapper.ProcessDefinitionMapper;
 import org.apache.dolphinscheduler.dao.mapper.ProjectMapper;
 import org.apache.dolphinscheduler.service.process.ProcessService;
+
+import java.text.ParseException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -36,13 +55,6 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import java.text.ParseException;
-import java.util.*;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.times;
-
 /**
  * test for ExecutorService
  */
@@ -50,7 +62,7 @@ import static org.mockito.Mockito.times;
 public class ExecutorService2Test {
 
     @InjectMocks
-    private ExecutorService executorService;
+    private ExecutorServiceImpl executorService;
 
     @Mock
     private ProcessService processService;
@@ -84,7 +96,7 @@ public class ExecutorService2Test {
     private String cronTime;
 
     @Before
-    public void init(){
+    public void init() {
         // user
         loginUser.setId(userId);
 
@@ -111,7 +123,6 @@ public class ExecutorService2Test {
 
     /**
      * not complement
-     * @throws ParseException
      */
     @Test
     public void testNoComplement() throws ParseException {
@@ -125,13 +136,12 @@ public class ExecutorService2Test {
                     Priority.LOW, Constants.DEFAULT_WORKER_GROUP, 110);
             Assert.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
             verify(processService, times(1)).createCommand(any(Command.class));
-        }catch (Exception e){
+        } catch (Exception e) {
         }
     }
 
     /**
      * date error
-     * @throws ParseException
      */
     @Test
     public void testDateError() throws ParseException {
@@ -145,13 +155,12 @@ public class ExecutorService2Test {
                     Priority.LOW, Constants.DEFAULT_WORKER_GROUP, 110);
             Assert.assertEquals(Status.START_PROCESS_INSTANCE_ERROR, result.get(Constants.STATUS));
             verify(processService, times(0)).createCommand(any(Command.class));
-        }catch (Exception e){
+        } catch (Exception e) {
         }
     }
 
     /**
      * serial
-     * @throws ParseException
      */
     @Test
     public void testSerial() throws ParseException {
@@ -165,17 +174,16 @@ public class ExecutorService2Test {
                     Priority.LOW, Constants.DEFAULT_WORKER_GROUP, 110);
             Assert.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
             verify(processService, times(1)).createCommand(any(Command.class));
-        }catch (Exception e){
+        } catch (Exception e) {
         }
     }
 
     /**
      * without schedule
-     * @throws ParseException
      */
     @Test
     public void testParallelWithOutSchedule() throws ParseException {
-        try{
+        try {
             Mockito.when(processService.queryReleaseSchedulerListByProcessDefinitionId(processDefinitionId)).thenReturn(zeroSchedulerList());
             Map<String, Object> result = executorService.execProcessInstance(loginUser, projectName,
                     processDefinitionId, cronTime, CommandType.COMPLEMENT_DATA,
@@ -185,17 +193,16 @@ public class ExecutorService2Test {
                     Priority.LOW, Constants.DEFAULT_WORKER_GROUP, 110);
             Assert.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
             verify(processService, times(31)).createCommand(any(Command.class));
-        }catch (Exception e){
+        } catch (Exception e) {
         }
     }
 
     /**
      * with schedule
-     * @throws ParseException
      */
     @Test
     public void testParallelWithSchedule() throws ParseException {
-        try{
+        try {
             Mockito.when(processService.queryReleaseSchedulerListByProcessDefinitionId(processDefinitionId)).thenReturn(oneSchedulerList());
             Map<String, Object> result = executorService.execProcessInstance(loginUser, projectName,
                     processDefinitionId, cronTime, CommandType.COMPLEMENT_DATA,
@@ -205,13 +212,13 @@ public class ExecutorService2Test {
                     Priority.LOW, Constants.DEFAULT_WORKER_GROUP, 110);
             Assert.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
             verify(processService, times(15)).createCommand(any(Command.class));
-        }catch (Exception e){
+        } catch (Exception e) {
         }
     }
 
 
     @Test
-    public void testNoMsterServers() throws ParseException{
+    public void testNoMsterServers() throws ParseException {
         Mockito.when(monitorService.getServerListFromZK(true)).thenReturn(new ArrayList<Server>());
 
         Map<String, Object> result = executorService.execProcessInstance(loginUser, projectName,
@@ -220,11 +227,11 @@ public class ExecutorService2Test {
                 null, null, 0,
                 "", "", RunMode.RUN_MODE_PARALLEL,
                 Priority.LOW, Constants.DEFAULT_WORKER_GROUP, 110);
-        Assert.assertEquals(result.get(Constants.STATUS),Status.MASTER_NOT_EXISTS);
+        Assert.assertEquals(result.get(Constants.STATUS), Status.MASTER_NOT_EXISTS);
 
     }
 
-    private List<Server> getMasterServersList(){
+    private List<Server> getMasterServersList() {
         List<Server> masterServerList = new ArrayList<>();
         Server masterServer1 = new Server();
         masterServer1.setId(1);
@@ -242,11 +249,11 @@ public class ExecutorService2Test {
 
     }
 
-    private List<Schedule> zeroSchedulerList(){
+    private List<Schedule> zeroSchedulerList() {
         return Collections.EMPTY_LIST;
     }
 
-    private List<Schedule> oneSchedulerList(){
+    private List<Schedule> oneSchedulerList() {
         List<Schedule> schedulerList = new LinkedList<>();
         Schedule schedule = new Schedule();
         schedule.setCrontab("0 0 0 1/2 * ?");
@@ -254,7 +261,7 @@ public class ExecutorService2Test {
         return schedulerList;
     }
 
-    private Map<String, Object> checkProjectAndAuth(){
+    private Map<String, Object> checkProjectAndAuth() {
         Map<String, Object> result = new HashMap<>();
         result.put(Constants.STATUS, Status.SUCCESS);
         return result;
